+++ src/components/pages/MissionRead/ContainerMissionRead.js
@@ -1,42 +1,36 @@
-import React from 'react';
-import Pdf from '@mikecousins/react-pdf';
+import React, { useEffect, useState } from 'react';
+import axios from 'axios';
+import { useLocalStorage } from '../../../utils/hooks';
 
-const RenderMissionRead = props => {
-  const { story, page, setPage } = props;
+import RenderMissionRead from './RenderMissionRead';
 
-  return (
-    <div className="pdf-container">
-      {story && (
-        <Pdf file={story} page={page}>
-          {({ pdfDocument, pdfPage, canvas }) => (
-            <>
-              {!pdfDocument && <span>Loading...</span>}
-              {canvas}
-              {Boolean(pdfDocument && pdfDocument.numPages) && (
-                <>
-                  <button
-                    className="prev-page-btn"
-                    disabled={page === 1}
-                    onClick={() => setPage(page - 1)}
-                  >
-                    {'<'}
-                  </button>
+const apiUrl = `${process.env.REACT_APP_API_URI}`;
 
-                  <button
-                    className="next-page-btn"
-                    disabled={page === pdfDocument.numPages}
-                    onClick={() => setPage(page + 1)}
-                  >
-                    {'>'}
-                  </button>
-                </>
-              )}
-            </>
-          )}
-        </Pdf>
-      )}
-    </div>
-  );
+const ContainerMissionRead = () => {
+  // URL to the PDF of the story to display
+  const [story, setStory] = useState('');
+
+  // holds the current page of the PDF to display
+  const [page, setPage] = useState(1);
+
+  const [curUserId, setCurUserId] = useLocalStorage('curUserId', null);
+  const [curUserToken, setCurUserToken] = useLocalStorage('curUserToken', null);
+  const id = curUserId;
+  const token = curUserToken;
+
+  useEffect(() => {
+    axios
+      .get(`${apiUrl}/child/${id}/mission`, {
+        headers: { Authorization: `Bearer ${token}` },
+      })
+      .then(res => {
+        console.log('stories', res.data);
+        setStory(res.data.read);
+      })
+      .catch(err => console.log({ err }));
+  }, [token, id]);
+
+  return <RenderMissionRead story={story} page={page} setPage={setPage} />;
 };
 
-export default RenderMissionRead;
+export default ContainerMissionRead;
